+++ src/clients/GitHub.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\authclient\clients;
@@ -32,7 +32,7 @@ use yii\authclient\OAuth2;
  * ]
  * ```
  *
- * @see http://developer.github.com/v3/oauth/
+ * @see https://developer.github.com/v3/oauth/
  * @see https://github.com/settings/developers
  *
  * @author Paul Klimov <klimov.paul@gmail.com>