+++ src/clients/Yandex.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\authclient\clients;
@@ -41,7 +41,7 @@ use yii\authclient\OAuth2;
  * ```
  *
  * @see https://oauth.yandex.ru/client/new
- * @see http://api.yandex.ru/login/doc/dg/reference/response.xml
+ * @see https://api.yandex.ru/login/doc/dg/reference/response.xml
  *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0