+++ src/signature/HmacSha.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\authclient\signature;
@@ -12,7 +12,7 @@ use yii\base\NotSupportedException;
 /**
  * HmacSha1 represents 'HMAC SHA' signature method.
  *
- * > **Note:** This class requires PHP "Hash" extension(<http://php.net/manual/en/book.hash.php>).
+ * > **Note:** This class requires PHP "Hash" extension(<https://php.net/manual/en/book.hash.php>).
  *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.1.3
@@ -21,7 +21,7 @@ class HmacSha extends BaseMethod
 {
     /**
      * @var string hash algorithm, e.g. `sha1`, `sha256` and so on.
-     * @see http://php.net/manual/ru/function.hash-algos.php
+     * @see https://php.net/manual/ru/function.hash-algos.php
      */
     public $algorithm;